-Construction Ghost Examine -Moved the ExamineFunc of the Construction to a shared helper
