+++ Content.Shared/GameObjects/EntitySystems/ConstructionSystem.cs
@@ -1,16 +1,22 @@
 ﻿using System;
+using Content.Shared.Construction;
 using JetBrains.Annotations;
 using Robust.Shared.GameObjects;
 using Robust.Shared.GameObjects.Systems;
+using Robust.Shared.IoC;
+using Robust.Shared.Localization;
 using Robust.Shared.Map;
 using Robust.Shared.Maths;
 using Robust.Shared.Serialization;
+using Robust.Shared.Utility;
 
 namespace Content.Shared.GameObjects.EntitySystems
 {
-    [UsedImplicitly]
     public class ConstructionSystem : EntitySystem
     {
+#pragma warning disable 649
+        [Dependency] private readonly ILocalizationManager _loc;
+#pragma warning restore 649
         /// <summary>
         ///     Sent client -> server to to tell the server that we started building
         ///     a structure-construction.
@@ -73,6 +79,29 @@ namespace Content.Shared.GameObjects.EntitySystems
             public AckStructureConstructionMessage(int ghostId)
             {
                 GhostId = ghostId;
+            }
+        }
+
+        public void DoExamine(FormattedMessage message, ConstructionPrototype prototype, int stage, bool inDetailRange)
+        {
+            var stages = prototype.Stages;
+            if (stage >= 0 && stage < stages.Count)
+            {
+                var curStage = stages[stage];
+                if (curStage.Backward != null && curStage.Backward is ConstructionStepTool)
+                {
+                    var backward = (ConstructionStepTool) curStage.Backward;
+                    message.AddText(_loc.GetString("To deconstruct: {0}x {1} Tool", backward.Amount, backward.ToolQuality));
+                }
+                if (curStage.Forward != null && curStage.Forward is ConstructionStepMaterial)
+                {
+                    if (curStage.Backward != null)
+                    {
+                        message.AddText("\n");
+                    }
+                    var forward = (ConstructionStepMaterial) curStage.Forward;
+                    message.AddText(_loc.GetString("To construct: {0}x {1}", forward.Amount, forward.Material));
+                }
             }
         }
     }
